backend/requirements.txt
django ~= 2.2.5
djangorestframework ~= 3.10.2
django-cors-headers ~= 3.1.0